✨: 게시물 목록을 표시하는 PostsTable 컴포넌트 추가 및 페이지에서 분리
